+++ docs/examples/03_backends_ros/files/04_plan_cartesian_motion.py
@@ -12,11 +12,11 @@ with RosClient() as client:
     frames.append(Frame([0.4, 0.3, 0.4], [0, 1, 0], [0, 0, 1]))
     start_configuration = Configuration.from_revolute_values([-0.042, 0.033, -2.174, 5.282, -1.528, 0.000])
 
-    response = robot.plan_cartesian_motion(frames,
-                                           start_configuration,
-                                           max_step=0.01,
-                                           avoid_collisions=True)
+    trajectory = robot.plan_cartesian_motion(frames,
+                                             start_configuration,
+                                             max_step=0.01,
+                                             avoid_collisions=True)
 
-    print("Computed cartesian path with %d configurations, " % len(response.configurations))
-    print("following %d%% of requested trajectory." % (response.fraction * 100))
-    print("Executing this path at full speed would take approx. %.3f seconds." % response.time_from_start)
+    print("Computed cartesian path with %d configurations, " % len(trajectory.points))
+    print("following %d%% of requested trajectory." % (trajectory.fraction * 100))
+    print("Executing this path at full speed would take approx. %.3f seconds." % trajectory.time_from_start)