+++ tests/test_sample.py
@@ -1,5 +1,0 @@
-import compas_fab
-
-
-def test_sample():
-    assert True