+++ CHANGELOG.rst
@@ -18,6 +18,10 @@ Unreleased
 * ROS Client: removed methods from ``Robot`` that are now handled with
   ``PlanningScene``, e.g. ``add_collision_mesh`` and
   ``add_attached_collision_mesh``
+* ROS Client: change the return type of ``plan_motion`` and ``plan_cartesian_motion``
+  to the new trajectory classes.
+* ROS File Server Loader: moved to ``compas_fab.backends`` package
+* ROS File Server Loader: renamed ``load`` to ``load_urdf`` and sync'd API to other loaders.
 * V-REP Client: renamed ``get_end_effector_pose`` to ``forward_kinematics``
 * V-REP Client: renamed ``find_robot_states`` to ``inverse_kinematics``
 * V-REP Client: renamed ``find_path_plan_to_config`` to
@@ -25,11 +29,14 @@ Unreleased
 * V-REP Client: renamed ``find_path_plan`` to ``plan_motion``
 * V-REP Client: changed ``is_connected`` to become a property
 * Made ``robot_artist`` default ``None`` on ``Robot`` constructor
+* Changed ``PathPlan`` class to use the new trajectory classes
 
 **Added**
 
+* Added ``scale`` method to ``Configuration``
 * Implemented Constraints (``OrientationConstraint``, ``PositionConstraint``, ``JointConstraint``) to use with ``plan_motion``
 * Implemented ``PlanningScene``, ``CollisionMesh`` and ``AttachedCollisionMesh``
+* Added generic representations for motion planning requests (``JointTrajectory``, ``JointTrajectoryPoint``, ``Duration``)
 * Added UR5 robot model data for example purposes
 * Added several doc examples